+++ src/test/java/racingcar/controller/RacingCarControllerTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.controller;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
